+++ ingenialink/dictionary.py
@@ -196,16 +196,16 @@ class Dictionary(ABC):
 
     version: str
     """Version of the dictionary."""
-    firmware_version: Optional[str]
+    firmware_version: Optional[str] = None
     """Firmware version declared in the dictionary."""
-    product_code: int
+    product_code: Optional[int] = None
     """Product code declared in the dictionary."""
     coco_product_code: Optional[int] = None
     """CoCo product code declared in the dictionary.
     Only used when a COM-KIT and a CORE dictionary are merged."""
-    part_number: Optional[str]
+    part_number: Optional[str] = None
     """Part number declared in the dictionary."""
-    revision_number: int
+    revision_number: Optional[int] = None
     """Revision number declared in the dictionary."""
     interface: Interface
     """Interface declared in the dictionary."""